fix: rename wsam directory references to wasm
