+++ wasm/swisseph.data
